% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerprof-package.R
\docType{package}
\name{cancerprof-package}
\alias{cancerprof}
\alias{cancerprof-package}
\title{cancerprof: API Client for State Cancer Profiles}
\description{
An interface for retrieving data from the NIH NCI State Cancer Profiles API \url{https://statecancerprofiles.cancer.gov/}. State Cancer Profiles provides information about data topics including demographics, screening and risk factors, cancer incidence, and mortality for US states, counties, and health service areas.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/getwilds/cancerprof}
  \item \url{https://getwilds.org/cancerprof/}
  \item Report bugs at \url{https://github.com/getwilds/cancerprof/issues}
}

}
\author{
\strong{Maintainer}: Brian Park \email{joon.brianpark@gmail.com} (\href{https://orcid.org/0009-0008-8274-3057}{ORCID})

}
\keyword{internal}
