% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-request.R
\name{create_request}
\alias{create_request}
\title{Create Request to State Cancer Profile Data}
\usage{
create_request(topic)
}
\arguments{
\item{topic}{Either "demographics", "risk", "incidencerates", "deathrates"}
}
\value{
returns the HTTP method with the state cancer profiles url and the http path
}
\description{
This creates a request to the State Cancer Profiles URL
}
\examples{
\dontrun{
create_request("demographics")
create_request("risk")
}
}
