% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-crowding.R
\name{demo_crowding}
\alias{demo_crowding}
\title{Access to Crowding Data}
\usage{
demo_crowding(area, areatype, crowding, race)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{crowding}{"household with >1 person per room"}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "white (includes hispanic)"
\item "white non-hispanic"
\item "black"
\item "amer. indian/alaskan native (includes hispanic)"
\item "asian or pacific islander (includes hispanic)"
\item "hispanic (any race)"
}}
}
\value{
A data frame with the following columns Area, Area Code, "Percent", "Households", "Rank"
}
\description{
This function returns a data frame containing demographic data on Crowding in State Cancer Profiles.
}
\examples{
\dontrun{
demo_crowding(area = "WA", 
              areatype = "hsa",
              crowding = "household with >1 person per room",
              race = "All Races (includes Hispanic)")
             
demo_crowding(area = "usa",
              areatype = "state",
              crowding = "household with >1 person per room",
              race = "All Races (includes Hispanic)")

demo_crowding(area = "pr",
              areatype = "hsa",
              crowding = "household with >1 person per room",
              race = "black")
}

}
