% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-insurance.R
\name{demo_insurance}
\alias{demo_insurance}
\title{Access to Insurance Data}
\usage{
demo_insurance(area, areatype, insurance, sex, age, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{insurance}{One of the following values:
\itemize{
\item \code{"\% Insured in demographic group, all income levels"}
\item \code{"\% Insured in demographic group, people at or below 138\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 200\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 250\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 400\% of Poverty"}
\item \code{"\% Insured in demographic group, people between 138\% - 400\% of poverty"}
\item \code{"\% uninsured in demographic group, all income levels"}
\item \code{"\% uninsured in demographic group, people at or below 138\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 200\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 250\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 400\% of Poverty"}
\item \code{"\% uninsured in demographic group, people between 138\% - 400\% of poverty"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}

\item{age}{If you specified \code{"both sexes"} for \code{sex} choose one of the following values:
\itemize{
\item \code{"under 19 years"}
\item \code{"18 to 64 years"}
\item \code{"21 to 64 years"}
\item \code{"40 to 64 years"}
\item \code{"50 to 64 years"}
\item \code{"under 65 years"}.
}

Otherwise if you specified \code{"male"} or \code{"female"} for \code{sex}, choose one of the following values:
\itemize{
\item \code{"18 to 64 years"}
\item \code{"40 to 64 years"}
\item \code{"50 to 64 years"}
\item \code{"Under 65 years"}.
}}

\item{race}{Only specify \code{race} if you specified \code{"state"} for \code{areatype}
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (non-Hispanic)"}
\item \code{"black (non-Hispanic)"}
\item \code{"American Indian / Alaska Native (non-Hispanic)"}
\item \code{"Asian (non-Hispanic)"}
\item \code{"Hispanic (Any Race)"}.
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, People, Rank.
}
\description{
This function returns a data frame about insurance demographics from State Cancer Profiles.
}
\examples{
\dontrun{
demo_insurance(
  area = "usa",
  areatype = "state",
  insurance = "\% Insured in demographic group, all income levels",
  sex = "both sexes",
  age = "18 to 64 years",
  race = "white (non-hispanic)"
)

demo_insurance(
  area = "wa",
  areatype = "hsa",
  insurance = "\% Insured in demographic group, all income levels",
  sex = "males",
  age = "18 to 64 years"
)

demo_insurance(
  area = "dc",
  areatype = "county",
  insurance = "\% Insured in demographic group, all income levels",
  sex = "males",
  age = "18 to 64 years"
)
}
}
