% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-mobility.R
\name{demo_mobility}
\alias{demo_mobility}
\title{Access to Mobility Data}
\usage{
demo_mobility(area, areatype, mobility)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{mobility}{One of five choices from "i haven't moved (in past year)", "moved from outside us (in past year)",
"moved, different state (in past year)", "moved, different county, same state (in past year)",
"moved, same county (in past year)"}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "People", "Rank"
}
\description{
This function returns a data frame from mobility in State Cancer Profiles
}
\examples{
\dontrun{
demo_mobility("WA", "county", "moved, different county, same state (in past year)")
demo_mobility("usa", "state", "moved, same county (in past year)")
demo_mobility("dc", "hsa", "moved, same county (in past year)")
}
}
