% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population.R
\name{demo_population}
\alias{demo_population}
\title{Access to Population Data}
\usage{
demo_population(area, areatype, population, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{population}{One of the following values:
\itemize{
\item "age under 18"
\item "age 18-39"
\item "age 40-64"
\item "ages 40 and over"
\item "ages 50 and over"
\item "ages 60 and over"
\item "american indian/alaska native"
\item "asian/pacific islander"
\item "black"
\item "foreign born"
\item "hispanic"
\item "non-hispanic (origin recode)"
\item "white"
\item "males"
\item "females"
}}

\item{race}{One of the following values:
\itemize{
\item "american indian/alaska native"
\item "asian/pacific islander"
\item "black"
\item "hispanic"
\item "white (includes hispanic)"
\item "white non-hispanic"
\item "hispanic (any race)"
}}

\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item Percent
\item Households
\item Rank
}
}
\description{
This function returns a data frame containing demographic data on Population in State Cancer Profiles.
}
\examples{
\dontrun{
demo_population(area = "WA",
                areatype = "county",
                population = "males",
                race = "all races (includes hispanic)")
                
demo_population(area = "dc",
                areatype = "hsa",
                population = "foreign born",
                race = "black",
                sex = "females")
                
demo_population(area = "usa",
                areatype = "state",
                population = "foreign born",
                race = "hispanic (any race)", 
                sex = "females")

}
}
