% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-poverty.R
\name{demo_poverty}
\alias{demo_poverty}
\title{Access to Poverty Data}
\usage{
demo_poverty(area, areatype, poverty, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"}, \code{"hsa"} (Health Service Area), or
\code{"state"}.}

\item{poverty}{The only permissible values are
\itemize{
\item \code{"families below poverty"}
\item \code{"persistent poverty"}
\item \code{"persons below poverty"}
\item \code{"persons < 150\% of poverty"}
}}

\item{race}{One of the following values
\itemize{
\item "All Races (includes Hispanic)"
\item "White (includes Hispanic)"
\item "White non-Hispanic"
\item "Black"
\item "Amer. Indian/Alaskan Native (includes Hispanic)"
\item "Asian or Pacific Islander (includes Hispanic)"
\item "hispanic (Any Race)
}}

\item{sex}{One of the following values
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Households, Rank
}
\description{
This function returns a data frame about poverty demographics from State Cancer Profiles.
}
\examples{
\dontrun{
demo_poverty(area = "WA",
             areatype = "county",
             poverty = "persistent poverty")
             
demo_poverty(area = "usa",
             areatype = "state",
             poverty = "families below poverty",
             race = "black")
             
demo_poverty(area = "dc",
             areatype = "hsa",
             poverty = "families below poverty",
             race = "All Races (includes Hispanic)")
}
}
