% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-svi.R
\name{demo_svi}
\alias{demo_svi}
\title{Access to Social Vulnerability Index (SVI) Data}
\usage{
demo_svi(area, svi)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{svi}{Either "Overall", "socioeconomic status", "household characteristics", "racial & ethinic minority status", "housing type & transportation"}
}
\value{
A data frame with the following columns "County", "FIPS", "Score"
}
\description{
This function returns a data frame from Social Vulnerability Index (SVI) in State Cancer Profiles
}
\examples{
\dontrun{
demo_svi("WA", "overall")
demo_svi("usa", "overall")
demo_svi("dc", "socioeconomic status")
}
}
