% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-dput-resp.R
\name{dput_resp}
\alias{dput_resp}
\title{Test dput-response}
\usage{
dput_resp()
}
\value{
A httr2 response object
}
\description{
This function creates a response object from State Cancer Profiles for
}
\examples{
\dontrun{
dput_resp()
}
}
