% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-dput-resp-incd.R
\name{dput_resp_incd}
\alias{dput_resp_incd}
\title{Test dput-resp-incidence}
\usage{
dput_resp_incd()
}
\value{
A httr2 response object
}
\description{
This function creates a response object from the Incidence section of State Cancer Profiles for
}
\examples{
\dontrun{
dput_resp_incd()
}
}
