% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-dput-resp-mortality.R
\name{dput_resp_mortality}
\alias{dput_resp_mortality}
\title{Test dput-resp-mortality}
\usage{
dput_resp_mortality()
}
\value{
A httr2 response object
}
\description{
This function creates a response object from the Mortality section of State Cancer Profiles for
}
\examples{
\dontrun{
dput_resp_mortality()
}
}
