% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-dput-resp-risk.R
\name{dput_resp_risk}
\alias{dput_resp_risk}
\title{Test dput-resp-demo}
\usage{
dput_resp_risk()
}
\value{
A httr2 response object
}
\description{
This function creates a response object from the Demographics section of State Cancer Profiles for
}
\examples{
\dontrun{
dput_resp_risk()
}
}
