% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips-scp.R
\name{fips_scp}
\alias{fips_scp}
\title{Process State to FIPS}
\usage{
fips_scp(area)
}
\arguments{
\item{area}{A state name or state abbreviation or United States of America or United States or usa or us}
}
\value{
A string fips value or NA
}
\description{
This function processes state names and abbreviations to FIPS data
}
\examples{
\dontrun{
fips_scp("pr")
fips_scp("ca")
fips_scp("usa")
}
}
