% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_area}
\alias{get_area}
\title{Get area titles and codes based on area type}
\usage{
get_area(areatype)
}
\arguments{
\item{areatype}{One of the following values:
\itemize{
\item "county"
\item "hsa"
\item "state".
}}
}
\value{
A character vector containing the title and code of the specified area type.
}
\description{
This function returns the title and code corresponding to the specified area type.
}
\examples{
 \dontrun{
get_area("county")
get_area("hsa")
}
}
