% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Metadata}
\usage{
get_metadata(input_tbl)
}
\arguments{
\item{input_tbl}{A tibble object}
}
\value{
a string of metadata and an invisible metadata object as a list
of strings
}
\description{
This function assigns a list of metadata components and returns a string of
processed metadata that is easily readable
}
\examples{
\dontrun{
process_metadata(resp)
}
}
