% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-age.R
\name{handle_age}
\alias{handle_age}
\title{Handles Age Values to Code}
\usage{
handle_age(age)
}
\arguments{
\item{sex}{Either "under 19 years", "18 to 64 years", "21 to 64 years", "40 to 64 years", "50 to 64 years", "under 65 years"}
}
\value{
A string for its respective age
}
\description{
This function returns a matching code value for a Age for the api to use to get data from State Cancer Profiles
}
\examples{
handle_age("under 19 years"), handle_age(age)
}
