% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-alcohol.R
\name{handle_alcohol}
\alias{handle_alcohol}
\title{Handles Women's Alcohol Values to Code}
\usage{
handle_alcohol(alcohol)
}
\arguments{
\item{alcohol}{binge drinking (4+ drinks on one occasion for women, 5+ drinks for one occasion for men), ages 21+}
}
\value{
A string for its respective Women's Health Value
}
\description{
This function returns a matching code value for Alcohol for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_alcohol(paste("binge drinking (4+ drinks on one occasion for women,"
                      "5+ drinks for one occasion for men), ages 21+"))
}
}
