% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-cancer.R
\name{handle_cancer}
\alias{handle_cancer}
\title{Handles Cancer Values to Code}
\usage{
handle_cancer(cancer)
}
\arguments{
\item{cancer}{Either "all cancer sites","bladder", "brain & ons", "breast (female)", "breast (female in situ)", "cervix",
"childhood (ages <15, all sites)", "childhood (ages <20, all sites)", "colon & rectum", "esophagus",
"kidney & renal pelvis", "leukemia", "liver & bile duct", "lung & bronchus", "melanoma of the skin",
"non-hodgkin lymphoma", "oral cavity & pharynx", "ovary", "pancreas", "prostate", "stomach",
"thyroid", "uterus (corpus & uterus, nos)"}
}
\value{
A string for its respective cancer type
}
\description{
This function returns a matching code value for a Cancer for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_cancer("bladder")
}
}
