% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-risk-datatype.R
\name{handle_datatype}
\alias{handle_datatype}
\title{Handles Risk Data Types to Code}
\usage{
handle_datatype(datatype)
}
\arguments{
\item{datatype}{Either "direct estimates" or "county level modeled estimates"}
}
\value{
A string for its respective Risk Data Types Value
}
\description{
This function returns a matching code value for a Risk Data Types for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_datatype("direct estimates")
}
}
