% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-education.R
\name{handle_education}
\alias{handle_education}
\title{Handles Education Values to Code}
\usage{
handle_education(education)
}
\arguments{
\item{education}{Either "less than 9th grade", "at least high school", "at least bachelors degree"}
}
\value{
A string for its respective education
}
\description{
This function returns a matching code value for a Education for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_education("at least bachelors degree")
}
}
