% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-food.R
\name{handle_food}
\alias{handle_food}
\title{Handles Food Values to Code}
\usage{
handle_food(food)
}
\arguments{
\item{food}{Either "food insecurity" or "limited access to healthy food"}
}
\value{
A string for its respective food
}
\description{
This function returns a matching code value for a Food for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_food("limited access to healthy food")
}
}
