% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income.R
\name{handle_income}
\alias{handle_income}
\title{Access to Income Data}
\usage{
handle_income(income)
}
\arguments{
\item{income}{Either "median family income" or "median household income"}

\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "HSA" (Health service area)}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}
}
\value{
A data frame with the following columns "County", "FIPS", "Dollars", "Rank"
}
\description{
This function returns a data frame from Income in State Cancer Profiles
}
\examples{
demo_income("wa", "county", "median family income", "all races (includes hispanic)")
}
