% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-insurance.R
\name{handle_insurance}
\alias{handle_insurance}
\title{Handles Insurance Values to Code}
\usage{
handle_insurance(insurance)
}
\arguments{
\item{insurance}{Either "\% insured in demographic group, all income levels" = "00030",
"\% insured in demographic group, people at or below 138\% of poverty",
"\% insured in demographic group, people at or below 200\% of poverty",
"\% insured in demographic group, people at or below 250\% of poverty",
"\% insured in demographic group, people at or below 400\% of poverty",
"\% insured in demographic group, people between 138\% - 400\% of poverty",
"\% uninsured in demographic group, all income levels",
"\% uninsured in demographic group, people at or below 138\% of poverty",
"\% uninsured in demographic group, people at or below 200\% of poverty",
"\% uninsured in demographic group, people at or below 250\% of poverty",
"\% uninsured in demographic group, people at or below 400\% of poverty",
"\% uninsured in demographic group, people between 138\% - 400\% of poverty"}
}
\value{
A string for its respective insurance
}
\description{
This function returns a matching code value for a insurance value for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_insurance("\% insured in demographic group, all income levels")
}
}
