% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobility.R
\name{handle_mobility}
\alias{handle_mobility}
\title{Access to Mobility Data}
\usage{
handle_mobility(mobility)
}
\arguments{
\item{mobility}{One of five choices from "i haven't moved (in past year)", "moved from outside us (in past year)",
"moved, different state (in past year)", "moved, different county, same state (in past year)",
"moved, same county (in past year)"}

\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "HSA" (Health service area)}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "People", "Rank"
}
\description{
This function returns a data frame from mobility in State Cancer Profiles
}
\examples{
demo_mobility("WA", "county", "moved, same county (in past year)")
}
