% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-population.R
\name{handle_population}
\alias{handle_population}
\title{Handles population Values to Code}
\usage{
handle_population(population)
}
\arguments{
\item{population}{One of the following values:
"age under 18", "age 18-39", age 40-64", "ages 40 and over", "ages 50 and over", "ages 60 and over",
"american indian/alaska native", "asian/pacific islander", "black", "foreign born",
"hispanic", "non-hispanic (origin recode)","white",

\if{html}{\out{<div class="sourceCode">}}\preformatted{                     "males", "females"
}\if{html}{\out{</div>}}}
}
\value{
A string for its respective population
}
\description{
This function returns a matching code value for a population for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_population("males")
}
}
