% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-poverty.R
\name{handle_poverty}
\alias{handle_poverty}
\title{Handles Poverty Values to Code}
\usage{
handle_poverty(poverty)
}
\arguments{
\item{poverty}{One of the following values:
"families below poverty", "persistent poverty", "persons below poverty", "persons < 150\% of poverty"}
}
\value{
A string for its respective poverty variable
}
\description{
This function returns a matching code value for a Poverty for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_poverty("families below poverty")
}
}
