% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-race.R
\name{handle_race}
\alias{handle_race}
\title{Handles Race Values to Code}
\usage{
handle_race(race)
}
\arguments{
\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","black (includes hispanic)", "amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}
}
\value{
A string for its respective race
}
\description{
This function returns a matching code value for a race for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_race("all races (includes hispanic)")
handle_race("black (non-hispanic)")
}
}
