% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-screening.R
\name{handle_screening}
\alias{handle_screening}
\title{Handles Colorectal Screening Values to Code}
\usage{
handle_screening(screening)
}
\arguments{
\item{screening}{Either "ever had colorectal endoscopy (sigmoidoscopy or colonoscopy), ages 50+",
"ever had fobt, ages 50-75", "fobt (1yr) / flex sig (5yr) / fobt (3yr) / colonoscopy (10yr), ages 50-75",
"guidance sufficient crc, ages 50-75", "had colonoscopy in past 10 years, ages 50-75",
"home-based fecal occult blood test (fobt) in past two years",
"home-based fobt in the past two years or ever had a colorectal endoscopy"}
}
\value{
A string for its respective Colorectal Screening Value
}
\description{
This function returns a matching code value for a Colorectal Screening for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_screening("guidance sufficient crc, ages 50-75")
}
}
