% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-sex.R
\name{handle_sex}
\alias{handle_sex}
\title{Handles Sex Values to Code}
\usage{
handle_sex(sex)
}
\arguments{
\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A string for its respective sex
}
\description{
This function returns a matching code value for a sex for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_sex("both sexes")
}
}
