% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-smoking.R
\name{handle_smoking}
\alias{handle_smoking}
\title{Handles Smoking Values to Code}
\usage{
handle_smoking(smoking)
}
\arguments{
\item{smoking}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65",
"pap smear in past 3 years, no hysterectomy, ages 18+"}
}
\value{
A string for its respective Smoking Value
}
\description{
This function returns a matching code value for a Smoking for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_smoking("former smoker, quit 1 year+; ages 18+")
}
}
