% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-stage.R
\name{handle_stage}
\alias{handle_stage}
\title{Handles Cancer Stage Values to Code}
\usage{
handle_stage(stage)
}
\arguments{
\item{stage}{Either "all stages" or "late stage (regional & distant)"}
}
\value{
A string for its respective Cancer Stage
}
\description{
This function returns a matching code value for a Cancer Stage for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_stage("all stages")
}
}
