% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-svi.R
\name{handle_svi}
\alias{handle_svi}
\title{Handles Social Vulnerability Index (SVI) Values to Code}
\usage{
handle_svi(svi)
}
\arguments{
\item{svi}{One of the following values:
"overall", "socioeconomic status", "household characteristics",
"racial & ethinic minority status", "housing type & transportation"}
}
\value{
A string for its respective SVI variable
}
\description{
This function returns a matching code value for a SVI for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_svi("overall")
}
}
