% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-vaccine.R
\name{handle_vaccine}
\alias{handle_vaccine}
\title{Handles Vaccines Values to Code}
\usage{
handle_vaccine(vaccine)
}
\arguments{
\item{vaccine}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65",
"pap smear in past 3 years, no hysterectomy, ages 18+"}
}
\value{
A string for its respective Vaccine Value
}
\description{
This function returns a matching code value for a Vaccines for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_vaccine("percent who received 2+ doses of HPV vaccine, ages 13-17")
}
}
