% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-whealth.R
\name{handle_whealth}
\alias{handle_whealth}
\title{Handles Women's Health Values to Code}
\usage{
handle_whealth(whealth)
}
\arguments{
\item{whealth}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65",
"pap smear in past 3 years, no hysterectomy, ages 18+"}
}
\value{
A string for its respective Women's Health Value
}
\description{
This function returns a matching code value for a Women's Health for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_whealth("mammogram in past 2 years, ages 50-74")
}
}
