% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-year.R
\name{handle_year}
\alias{handle_year}
\title{Handles Cancer Incidence Year Values to Code}
\usage{
handle_year(year)
}
\arguments{
\item{year}{Either "latest 5 year average", "latest single year (us by state)"}
}
\value{
A string for its respective Cancer Incidence Year
}
\description{
This function returns a matching code value for a Cancer Incidence Year for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_year("latest 5 year average")
}
}
