% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality-cancer.R
\name{mortality_cancer}
\alias{mortality_cancer}
\title{Access to Cancer Mortality Data}
\usage{
mortality_cancer(area, areatype, cancer, race, sex, age, year)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{cancer}{One of the following values:
\itemize{
\item \code{"all cancer sites"}
\item \code{"bladder"}
\item \code{"brain & ons"}
\item \code{"breast (female)"}
\item \code{"cervix"}
\item \code{"childhood (ages <15, all sites)"}
\item \code{"childhood (ages <20, all sites)"}
\item \code{"colon & rectum"}
\item \code{"esophagus"}
\item \code{"kidney & renal pelvis"}
\item \code{"leukemia"}
\item \code{"liver & bile duct"}
\item \code{"lung & bronchus"}
\item \code{"melanoma of the skin"}
\item \code{"non-hodgkin lymphoma"}
\item \code{"oral cavity & pharynx"}
\item \code{"ovary"}
\item \code{"pancreas"}
\item \code{"prostate"}
\item \code{"stomach"}
\item \code{"thyroid"}
\item \code{"uterus (corpus & uterus, nos)"}
}}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (non-Hispanic)"}
\item \code{"Black (non-Hispanic)"}
\item \code{"American Indian / Alaska Native (non-Hispanic)"}
\item \code{"Asian / Pacific Islander (non-Hispanic)"}
\item \code{"Hispanic (Any Race)"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}

\item{age}{One of the following values:
\itemize{
\item \code{"all ages"}
\item \code{"ages <50"}
\item \code{"ages 50+"}
\item \code{"ages <65"}
\item \code{"ages 65+"}
\item \code{"ages <15"}
\item \code{"ages <20"}.
}}

\item{year}{One of the following values:
\itemize{
\item \code{"latest 5 year average"}
\item \code{"latest single year (us by state)"}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Met Healthy People Objective of ***?,
Age Adjusted Death Rate, Lower 95\% CI Rate, Upper 95\% CI Rate,
CI Rank, Lower CI Rank, Upper CI Rank, Annual Average Count,
Recent Trend, Recent 5 Year Trend, Lower 95\% CI Trend, Upper 95\% CI Trend.
}
\description{
This function returns a data frame about cancer mortality
from State Cancer Profiles.
}
\examples{
\dontrun{
mortality_cancer(
  area = "wa",
  areatype = "county",
  cancer = "all cancer sites",
  race = "black (non-hispanic)",
  sex = "both sexes",
  age = "ages 65+",
  year = "latest 5 year average"
)

mortality_cancer(
  area = "usa",
  areatype = "state",
  cancer = "prostate",
  race = "all races (includes hispanic)",
  sex = "males",
  age = "ages 50+",
  year = "latest single year (us by state)"
)

mortality_cancer(
  area = "wa",
  areatype = "hsa",
  cancer = "ovary",
  race = "all races (includes hispanic)",
  sex = "females",
  age = "ages 50+",
  year = "latest 5 year average"
)
}
}
