% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-incidence.R
\name{process_incidence}
\alias{process_incidence}
\title{Process Cancer Incidence Response Data}
\usage{
process_incidence(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A processed response data frame
}
\description{
This function processes the Cancer Incidence response data from State Cancer Profiles
}
\examples{
\dontrun{
process_incidence(resp)
}
}
