% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-metadata.R, R/process_metadata.R
\name{process_metadata}
\alias{process_metadata}
\title{Process Metadata}
\usage{
process_metadata(resp)

process_metadata(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A response object with Metadata and a tibble

A response object with Metadata and a tibble
}
\description{
This function sets the class of the response data
to use the custom print function

This function sets the class of the response data
to use the custom print function
}
\examples{
\dontrun{
process_metadata(resp)
}
\dontrun{
process_metadata(resp)
}
}
