% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-mortality.R
\name{process_mortality}
\alias{process_mortality}
\title{Process Cancer Mortality Response Data}
\usage{
process_mortality(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A processed response data frame
}
\description{
This function processes the Cancer Mortality response data from State Cancer Profiles
}
\examples{
\dontrun{
process_mortality(resp)
}
}
