% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-response.R
\name{process_response}
\alias{process_response}
\title{Process Response Data}
\usage{
process_response(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A processed response data frame
}
\description{
This function processes the response data from State Cancer Profiles
}
\examples{
process_response(resp)
}
