% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-screening.R
\name{process_screening}
\alias{process_screening}
\title{Process Response Data}
\usage{
process_screening(resp)
}
\arguments{
\item{resp}{A response object}
}
\value{
A processed response data frame
}
\description{
This function processes the response data from State Cancer Profiles
}
\examples{
\dontrun{
process_screening(resp)
}
}
