% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull-trend-data.R
\name{pull_trend_data}
\alias{pull_trend_data}
\title{Access to Trend Data}
\usage{
pull_trend_data(area, cancer, race, sex, age, datatype)
}
\arguments{
\item{area}{A state/territory abbreviation or USA or "seer 9 registeries"}

\item{cancer}{One of the following values:
\itemize{
\item \code{"all cancer sites"}
\item \code{"bladder"}
\item \code{"brain & ons"}
\item \code{"breast (female)"}
\item \code{"breast (female in situ)"}
\item \code{"cervix"}
\item \code{"childhood (ages <15, all sites)"}
\item \code{"childhood (ages <20, all sites)"}
\item \code{"colon & rectum"}
\item \code{"esophagus"}
\item \code{"kidney & renal pelvis"}
\item \code{"leukemia"}
\item \code{"liver & bile duct"}
\item \code{"lung & bronchus"}
\item \code{"melanoma of the skin"}
\item \code{"non-hodgkin lymphoma"}
\item \code{"oral cavity & pharynx"}
\item \code{"ovary"}
\item \code{"pancreas"}
\item \code{"prostate"}
\item \code{"stomach"}
\item \code{"thyroid"}
\item \code{"uterus (corpus & uterus, nos)"}.
}}

\item{race}{One of the following values:
\itemize{
\item \code{"all Races (includes Hispanic)"}
\item \code{"white non-hispanic"}
\item \code{"black (non-hispanic)"}
\item \code{"american indian / alaska native (non-hispanic)"}
\item \code{"asian / pacific islander (non-hispanic)"}
\item \code{"hispanic (Any Race)"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}

\item{age}{One of the following values:
\itemize{
\item \code{"all ages"}
\item \code{"ages <50"}
\item \code{"ages 50+"}
\item \code{"ages <65"}
\item \code{"ages 65+"}
}}

\item{datatype}{One of the following values:
\itemize{
\item \code{"incidence"}
\item \code{"mortality"}
}}
}
\value{
A data frame with the following columns: Year, Observed, Estimated
}
\description{
This function returns a data frame about trend data on incidence and
mortality from State Cancer Profiles
}
\examples{
\dontrun{
pull_trend_data(
  area = "wa",
  cancer = "lung & bronchus",
  race = "All Races (includes Hispanic)",
  sex = "both sexes",
  age = "all ages",
  datatype = "incidence"
)
# no data available
pull_trend_data(
  area = "wa",
  cancer = "lung & bronchus",
  race = "black (non-hispanic)",
  sex = "females",
  age = "ages 65+",
  datatype = "incidence"
)
}
}
