% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-smoking.R
\name{risk_smoking}
\alias{risk_smoking}
\title{Access to Smoking Data}
\usage{
risk_smoking(smoking, race = NULL, sex = NULL, datatype = NULL, area = NULL)
}
\arguments{
\item{smoking}{One of the following values:
\itemize{
\item "smoking laws (any)"
\item "smoking laws (bars)"
\item "smoking laws (restaurants)"
\item "smoking laws (workplace)"
\item "smoking laws (workplace; restaurant; & bar)"
\item "smokers (stopped for 1 day or longer)"
\item "smoking not allowed at work (all people)"
\item "smoking not allowed in home (all people)"
\item "smoking not allowed at work (current smokers)"
\item "smoking not allowed at work (former/never smokers)"
\item "smoking not allowed in home (current smokers)"
\item "smoking not allowed in home (former/never smokers)"
\item "former smoker; ages 18+"
\item "former smoker, quit 1 year+; ages 18+"
\item "smokers (ever); ages 18+"
\item "e-cigarette use; ages 18+"
\item "smokers (current); ages 18+"
}}

\item{race}{One of the following values:
\itemize{
\item "all races (includes hispanic)"
\item "white (non-hispanic)"
\item "black (non-hispanic)"
\item "amer. indian / ak native (non-hispanic)"
\item "asian / pacific islander (non-hispanic)"
\item "hispanic (any race)"
}}

\item{sex}{Either "both sexes", "males", "females"}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item "Percent"
\item "Lower CI 95\%"
\item "Upper CI 95\%"
\item "Number of Respondents"
}
}
\description{
This function returns a data frame that includes data on screening and risk
factors related to smoking as presented in the State Cancer Profiles.
}
\details{
Please note that this function requires very specific arguments for each smoking type
}
\examples{
\dontrun{
risk_smoking(smoking = "smoking laws (any)")

risk_smoking(smoking = "smokers (stopped for 1 day or longer)",
             sex = "both sexes", 
             datatype = "county level modeled estimates",
             area = "wa")
             
risk_smoking(smoking = "smoking not allowed at work (current smokers)",
             sex = "both sexes", 
             datatype = "direct estimates")
             
risk_smoking(smoking = "smokers (current); ages 18+",
             race = "all races (includes hispanic)",
             sex = "both sexes",
             datatype = "county level modeled estimates",
             area="wa")
}
}
