% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-vaccines.R
\name{risk_vaccine}
\alias{risk_vaccine}
\title{Access to Vaccines Data}
\usage{
risk_vaccine(vaccine, sex)
}
\arguments{
\item{vaccine}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65",
"pap smear in past 3 years, no hysterectomy, ages 18+"}

\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A data frame with the following columns "State", "FIPS", "Percent", "Lower 95\% CI", "Upper 95\% CI", "Number of Respondents"
}
\description{
This function returns a data frame from Vaccines in State Cancer Profiles
}
\examples{
\dontrun{
risk_vaccine("percent who received 2+ doses of HPV vaccine, ages 13-15", "both sexes")
risk_vaccine("percent who received 2+ doses of HPV vaccine, ages 13-17", "both sexes")
risk_vaccine("percent who received 3+ doses of HPV vaccine, ages 13-17", "females")
}
}
