% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-vaccines.R
\name{risk_vaccines}
\alias{risk_vaccines}
\title{Access to Vaccines Data}
\usage{
risk_vaccines(vaccine, sex)
}
\arguments{
\item{vaccine}{Either "percent with up to date hpv vaccination coverage, ages 13-15",
"percent with up to date hpv vaccination coverage, ages 13-17"}

\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "Lower 95\% CI", "Upper 95\% CI", "Number of Respondents"
}
\description{
This function returns a data frame from Vaccines in State Cancer Profiles
}
\examples{
\dontrun{
risk_vaccines(vaccine = "percent with up to date hpv vaccination coverage, ages 13-15",
              sex = "both sexes")
              
risk_vaccines(vaccine = "percent with up to date hpv vaccination coverage, ages 13-17",
              sex = "females")
}
}
