% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-womens-health.R
\name{risk_whealth}
\alias{risk_whealth}
\title{Access to Women's Health Data}
\usage{
risk_whealth(whealth, race, datatype = "direct estimates", area = NULL)
}
\arguments{
\item{whealth}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Women's Health in State Cancer Profiles
}
\examples{
\dontrun{
risk_whealth("mammogram in past 2 years, ages 50-74", 
             "all races (includes hispanic)", "direct estimates")
risk_whealth("pap smear in past 3 years, no hysterectomy, ages 21-65", 
             "all races (includes hispanic)", "county level modeled estimates", "wa")
risk_whealth("pap smear in past 3 years, no hysteroetomy, ages 21-65", "black (non-hispanic)")
}
}
