% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-womens-health.R
\name{risk_women_health}
\alias{risk_women_health}
\title{Access to Women's Health Data}
\usage{
risk_women_health(
  women_health,
  race,
  datatype = "direct estimates",
  area = NULL
)
}
\arguments{
\item{women_health}{Either "mammogram in past 2 years, ages 50-74", "mammogram in past 2 years, ages 40+",
"pap smear in past 3 years, no hysterectomy, ages 21-65"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Women's Health in State Cancer Profiles
}
\examples{
\dontrun{
risk_women_health(women_health = "mammogram in past 2 years, ages 50-74", 
                  race = "all races (includes hispanic)",
                  datatype = "direct estimates")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysterectomy, ages 21-65", 
                  race = "all races (includes hispanic)",
                  datatype = "county level modeled estimates",
                  area = "wa")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysteroetomy, ages 21-65",
                  race = "black (non-hispanic)")
}
}
