% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-crowding.R
\docType{data}
\name{area}
\alias{area}
\title{Access to Crowding Data}
\format{
An object of class \code{character} of length 1.
}
\usage{
area
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{race}{One of the following values:
"All Races (includes Hispanic)",
"white (includes hispanic)",
"white non-hispanic",
"black",
"amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)",
"hispanic (any race)"}
}
\value{
A data frame with the following columns Area, Area Code, "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from Crowding in State Cancer Profiles
}
\examples{
\dontrun{
demo_crowding(area = "WA", 
              areatype = "hsa", 
              race = "All Races (includes Hispanic)")
             
demo_crowding(area = "usa",
              areatype = "state",
              race = "All Races (includes Hispanic)")

demo_crowding(area = "pr",
              areatype = "hsa",
              race = "black")
}

}
\keyword{datasets}
