% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerprof-package.R
\docType{package}
\name{cancerprof-package}
\alias{cancerprof}
\alias{cancerprof-package}
\title{cancerprof: API Client for State Cancer Profiles}
\description{
An API Client for accessing data from State Cancer Profiles for programmable analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/getwilds/cancerprof}
  \item \url{http://getwilds.org/cancerprof/}
  \item Report bugs at \url{https://github.com/getwilds/cancerprof/issues}
}

}
\author{
\strong{Maintainer}: Brian Park \email{joon.brianpark@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
