% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerprof-package.R
\docType{package}
\name{cancerprof-package}
\alias{cancerprof}
\alias{cancerprof-package}
\title{cancerprof: Open Cancer Informatics}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Brian Park \email{joon.brianpark@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
