% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crowding.R
\name{demo_crowding}
\alias{demo_crowding}
\title{Access to Crowding Data}
\usage{
demo_crowding(area, areatype, race)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "HSA" (Health service area)}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from Crowding in State Cancer Profiles
}
\examples{
\dontrun{
demo_crowding("WA", "county", "All Races (includes Hispanic)")
}
}
