% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-education.R
\name{demo_education}
\alias{demo_education}
\title{Access to Education Data}
\usage{
demo_education(area, areatype, education, sex = NULL, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"}, \code{"hsa"} (Health Service Area), or
\code{"state"}.}

\item{education}{The only permissible values are
\itemize{
\item \code{"less than 9th grade"}
\item \code{"at least high school"}
\item \code{"at least bachelors degree"}
}}

\item{sex}{One of the following values:
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "White (includes Hispanic)"
\item "White non-Hispanic"
\item "Black"
\item "Amer. Indian/Alaskan Native (includes Hispanic)"
\item "Asian or Pacific Islander (includes Hispanic)"
\item "hispanic (Any Race)
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Households, Rank
}
\description{
This function returns a data frame about education demographics from State Cancer Profiles.
}
\examples{
\dontrun{
demo_education(area = "wa",
               areatype = "county",
               education = "at least high school",
               sex = "males")
               
demo_education(area = "usa",
               areatype = "state",
               education = "at least bachelors degree",
               sex = "both sexes",
               race = "all races (includes hispanic)")
               
demo_education(area = "pr",
               areatype = "hsa",
               education = "less than 9th grade")
}
}
