% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-education.R
\name{demo_education}
\alias{demo_education}
\title{Access to Education Data}
\usage{
demo_education(area, areatype, education, sex = NULL, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{education}{One of the following values:
\itemize{
\item \code{"less than 9th grade"}
\item \code{"at least high school"}
\item \code{"at least bachelors degree"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (includes Hispanic)"}
\item \code{"White non-Hispanic"}
\item \code{"Black"}
\item \code{"Amer. Indian/Alaskan Native (includes Hispanic)"}
\item \code{"Asian or Pacific Islander (includes Hispanic)"}
\item \verb{"Hispanic (Any Race)}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Percent, Households, Rank.
}
\description{
This function returns a data frame about education demographics from
State Cancer Profiles.
}
\examples{
\dontrun{
demo_education(
  area = "wa",
  areatype = "county",
  education = "at least high school",
  sex = "males"
)

demo_education(
  area = "usa",
  areatype = "state",
  education = "at least bachelors degree",
  sex = "both sexes",
  race = "all races (includes hispanic)"
)

demo_education(
  area = "pr",
  areatype = "hsa",
  education = "less than 9th grade"
)
}
}
\seealso{
Other demographics: 
\code{\link{demo_crowding}()},
\code{\link{demo_food}()},
\code{\link{demo_income}()},
\code{\link{demo_insurance}()},
\code{\link{demo_language}()},
\code{\link{demo_mobility}()},
\code{\link{demo_population}()},
\code{\link{demo_poverty}()},
\code{\link{demo_svi}()},
\code{\link{demo_workforce}()}
}
\concept{demographics}
