% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-education.R
\name{demo_education}
\alias{demo_education}
\title{Access to Education Data}
\usage{
demo_education(area, areatype, education, sex = NULL, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{education}{Either "less than 9th grade", "at least high school", "at least bachelors degree"}

\item{sex}{Either "both sexes", "male", "female"}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "white (includes hispanic)"
\item "white non-hispanic"
\item "black"
\item "amer. indian/alaskan native (includes hispanic)"
\item "asian or pacific islander (includes hispanic)"
\item "hispanic (any race)
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "Households", "Rank"
}
\description{
This function returns a data frame containing demographic data on Education in State Cancer Profiles.
}
\examples{
\dontrun{
demo_education(area = "wa",
               areatype = "county",
               education = "at least high school",
               sex = "males")
               
demo_education(area = "usa",
               areatype = "state",
               education = "at least bachelors degree",
               sex = "both sexes",
               race = "all races (includes hispanic)")
               
demo_education(area = "pr",
               areatype = "hsa",
               education = "less than 9th grade")
}
}
