% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-food-access.R
\name{demo_food}
\alias{demo_food}
\title{Access to Food Insecurity Data}
\usage{
demo_food(area, areatype, food, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "state"}

\item{food}{Either "food insecurity" or "limited access to healthy food"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white non hispanic",
"black (includes hispanic)","hispanic (any race)}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Value", "People"
}
\description{
This function returns a data frame from Food Insecurity in State Cancer Profiles
}
\examples{
\dontrun{
demo_food(area = "wa",
          areatype = "county",
          food = "food insecurity", 
          race = "black")
          
demo_food(area = "usa",
          areatype = "state",
          food = "limited access to healthy food")
          
demo_food(area = "pr",
          areatype = "county",
          food = "food insecurity",
          race = "all races (includes hispanic)")
}
}
