% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-food-access.R
\name{demo_food}
\alias{demo_food}
\title{Access to Food Insecurity Data}
\usage{
demo_food(area, areatype, food, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "state"}

\item{food}{Either "food insecurity" or "limited access to healthy food"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white non hispanic" = "01",
"black (includes hispanic)","hispanic (any race)}
}
\value{
A data frame with the following columns "County", "FIPS", "Value", "People"
}
\description{
This function returns a data frame from Food Insecurity in State Cancer Profiles
}
\examples{
\dontrun{
demo_food("wa", "county", "food insecurity", "black")
demo_food("usa", "state", "limited access to healthy food")
demo_food("pr", "county", "food insecurity", "all races (includes hispanic)")
}
}
