% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-income.R
\name{demo_income}
\alias{demo_income}
\title{Access to Income Data}
\usage{
demo_income(area, areatype, income, race)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "state"}

\item{income}{Either "median family income" or "median household income"}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "white (includes hispanic)"
\item "white non-hispanic"
\item "black",
\item "amer. indian/alaskan native (includes hispanic)"
\item "asian or pacific islander (includes hispanic)"
\item "hispanic (any race)
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Dollars", "Rank"
}
\description{
This function returns a data frame containing demographic data on Income in State Cancer Profiles
}
\examples{
\dontrun{
demo_income(area = "wa",
            areatype = "county",
            income = "median household income",
            race = "all races (includes hispanic)")
            
demo_income(area = "usa",
            areatype = "state",
            income = "median family income",
            race = "all races (includes hispanic)")
            
demo_income(area = "pr",
            areatype = "county",
            income = "median family income",
            race = "all races (includes hispanic)")
}
}
