% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-income.R
\name{demo_income}
\alias{demo_income}
\title{Access to Income Data}
\usage{
demo_income(area, areatype, income, race)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"} or \code{"state"}.}

\item{income}{Either \code{"median family income"} or \code{"median household income"}.}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (includes Hispanic)"}
\item \code{"White non-Hispanic"}
\item \code{"Black"}
\item \code{"Amer. Indian/Alaskan Native (includes Hispanic)"}
\item \code{"Asian or Pacific Islander (includes Hispanic)"}
\item \verb{"Hispanic (Any Race)}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Dollars, Rank.
}
\description{
This function returns a data frame about income demographics from
State Cancer Profiles.
}
\examples{
\dontrun{
demo_income(
  area = "wa",
  areatype = "county",
  income = "median household income",
  race = "all races (includes hispanic)"
)

demo_income(
  area = "usa",
  areatype = "state",
  income = "median family income",
  race = "all races (includes hispanic)"
)

demo_income(
  area = "pr",
  areatype = "county",
  income = "median family income",
  race = "all races (includes hispanic)"
)
}
}
\seealso{
Other demographics: 
\code{\link{demo_crowding}()},
\code{\link{demo_education}()},
\code{\link{demo_food}()},
\code{\link{demo_insurance}()},
\code{\link{demo_language}()},
\code{\link{demo_mobility}()},
\code{\link{demo_population}()},
\code{\link{demo_poverty}()},
\code{\link{demo_svi}()},
\code{\link{demo_workforce}()}
}
\concept{demographics}
