% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-insurance.R
\name{demo_insurance}
\alias{demo_insurance}
\title{Access to Insurance Data}
\usage{
demo_insurance(area, areatype, insurance, sex, age, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{insurance}{Either
"\% Insured in demographic group, all income levels", "\% Insured in demographic group, people at or below 138\% of Poverty"
"\% Insured in demographic group, people at or below 200\% of Poverty", "\% Insured in demographic group, people at or below 250\% of Poverty"
"\% Insured in demographic group, people at or below 400\% of Poverty","\% Insured in demographic group, people between 138\% - 400\% of poverty"
"\% uninsured in demographic group, all income levels","\% uninsured in demographic group, people at or below 138\% of Poverty"
"\% uninsured in demographic group, people at or below 200\% of Poverty","\% uninsured in demographic group, people at or below 250\% of Poverty"
"\% uninsured in demographic group, people at or below 400\% of Poverty","\% uninsured in demographic group, people between 138\% - 400\% of poverty"}

\item{sex}{Either "both sexes", "male", "female"}

\item{age}{Either "under 19 years", "18 to 64 years","21 to 64 years","40 to 64 years","50 to 64 years","under 65 years" for "both sexes"
"18 to 64 years","40 to 64 years","50 to 64 years","Under 65 years" for "males" and "females"}

\item{race}{Either "all races (includes hispanic)", "white (non-hispanic)", "black (non-hispanic)", "american indian / alaska native non-hispanic",
"asian (non-hispanic)", "hispanic (any race)"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "People", "Rank"
}
\description{
This function returns a data frame from Insurance in State Cancer Profiles
}
\examples{
\dontrun{
demo_insurance(area = "usa",
               areatype = "state",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "both sexes",
               age = "18 to 64 years",
               race = "white (non-hispanic)")
               
demo_insurance(area = "wa", 
               areatype = "hsa",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "males",
               age = "18 to 64 years")
               
demo_insurance(area = "dc",
               areatype = "county",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "males",
               age = "18 to 64 years")
}
}
