% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-insurance.R
\name{demo_insurance}
\alias{demo_insurance}
\title{Access to Insurance Data}
\usage{
demo_insurance(area, areatype, insurance, sex, age, race = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"}, \code{"hsa"} (Health Service Area), or
\code{"state"}.}

\item{insurance}{The only permissible values are:
\itemize{
\item \code{"\% Insured in demographic group, all income levels"}
\item \code{"\% Insured in demographic group, people at or below 138\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 200\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 250\% of Poverty"}
\item \code{"\% Insured in demographic group, people at or below 400\% of Poverty"}
\item \code{"\% Insured in demographic group, people between 138\% - 400\% of poverty"}
\item \code{"\% uninsured in demographic group, all income levels"}
\item \code{"\% uninsured in demographic group, people at or below 138\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 200\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 250\% of Poverty"}
\item \code{"\% uninsured in demographic group, people at or below 400\% of Poverty"}
\item \code{"\% uninsured in demographic group, people between 138\% - 400\% of poverty"}
}}

\item{sex}{Either "both sexes", "male", "female"}

\item{age}{The only permissible values for "both sexes" are:
\itemize{
\item "under 19 years"
\item "18 to 64 years"
\item "21 to 64 years"
\item "40 to 64 years"
\item "50 to 64 years"
\item "under 65 years"
}

The only permissible values for "males" and "females" are:
\itemize{
\item "18 to 64 years"
\item "40 to 64 years"
\item "50 to 64 years"
\item "Under 65 years"
}}

\item{race}{One of the following values:
\itemize{
\item All Races (includes Hispanic)"
\item "White (non-Hispanic)"
\item "black (non-Hispanic)"
\item "American Indian / Alaska Native (non-Hispanic)"
\item "Asian (non-Hispanic)"
\item "Hispanic (Any Race)"
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "People", "Rank"
}
\description{
This function returns a data frame about insurance demographics from State Cancer Profiles.
}
\examples{
\dontrun{
demo_insurance(area = "usa",
               areatype = "state",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "both sexes",
               age = "18 to 64 years",
               race = "white (non-hispanic)")
               
demo_insurance(area = "wa", 
               areatype = "hsa",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "males",
               age = "18 to 64 years")
               
demo_insurance(area = "dc",
               areatype = "county",
               insurance = "\% Insured in demographic group, all income levels", 
               sex = "males",
               age = "18 to 64 years")
}
}
