% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-non-english-language.R
\name{demo_language}
\alias{demo_language}
\title{Access to Non-English Language}
\usage{
demo_language(area, areatype, language)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{language}{"language isolation"}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item "Percent"
\item "Households"
\item "Rank"
}
}
\description{
This function returns a data frame containing demographic data on Non-English Language in State Cancer Profiles.
}
\examples{
demo_language(area = "WA",
              areatype = "county",
              language = "language isolation")
              
demo_language(area = "dc",
              areatype = "hsa",
              language = "language isolation")
              
demo_language(area = "usa", 
              areatype = "state",
              language = "language isolation")
}
