% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-non-english-language.R
\name{demo_language}
\alias{demo_language}
\title{Access to Non-English Language}
\usage{
demo_language(area, areatype)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from Crowding in State Cancer Profiles
}
\examples{
demo_language("WA", "county")
demo_language("dc", "hsa")
demo_language("usa", "state")
}
