% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-mobility.R
\name{demo_mobility}
\alias{demo_mobility}
\title{Access to Mobility Data}
\usage{
demo_mobility(area, areatype, mobility)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{mobility}{The only permissible values are
\itemize{
\item \code{"i haven't moved (in past year)"}
\item \code{"moved from outside us (in past year)"}
\item \code{"moved, different state (in past year)"}
\item \code{"moved, different county, same state (in past year)"}
\item \code{"moved, same county (in past year)"}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Percent, People, Rank.
}
\description{
This function returns a data frame about mobility demographics
from State Cancer Profiles.
}
\examples{
\dontrun{
demo_mobility(
  area = "WA",
  areatype = "county",
  mobility = "moved, different county, same state (in past year)"
)

demo_mobility(
  area = "usa",
  areatype = "state",
  mobility = "moved, same county (in past year)"
)

demo_mobility(
  area = "dc",
  areatype = "hsa",
  mobility = "moved, same county (in past year)"
)
}
}
