% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population.R
\name{demo_population}
\alias{demo_population}
\title{Access to Population Data}
\usage{
demo_population(area, areatype, population, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{population}{One of the following values: "age under 18", "age 18-39", "age 40-64", "ages 40 and over", "ages 50 and over", "ages 60 and over",
"american indian/alaska native", "asian/pacific islander", "black", "foreign born", "hispanic",
"non-hispanic (origin recode)", "white", "males", "females"}

\item{race}{One of the following values: "american indian/alaska native", "asian/pacific islander",
"black", "hispanic", "white (includes hispanic)", "white non-hispanic", "hispanic (any race)"}

\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from population in State Cancer Profiles
}
\examples{
\dontrun{
demo_population(area = "WA",
                areatype = "county",
                population = "males",
                race = "all races (includes hispanic)")
                
demo_population(area = "dc",
                areatype = "hsa",
                population = "foreign born",
                race = "black",
                sex = "females")
                
demo_population(area = "usa",
                areatype = "state",
                population = "foreign born",
                race = "hispanic (any race)", 
                sex = "females")

}
}
