% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population.R
\name{demo_population}
\alias{demo_population}
\title{Access to Population Data}
\usage{
demo_population(area, areatype, population, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"}, \code{"hsa"} (Health Service Area), or
\code{"state"}.}

\item{population}{The only permissible values are
\itemize{
\item \code{"age under 18"}
\item \code{"age 18-39"}
\item \code{"age 40-64"}
\item \code{"ages 40 and over"}
\item \code{"ages 50 and over"}
\item \code{"ages 60 and over"}
\item \code{"american indian/alaska native"}
\item \code{"asian/pacific islander"}
\item \code{"black"}
\item \code{"foreign born"}
\item \code{"hispanic"}
\item \code{"non-hispanic (origin recode)"}
\item \code{"white"}
\item \code{"males"}
\item \code{"females"}
}}

\item{race}{One of the following values
\itemize{
\item "American Indian/Alaska Native"
\item "Asian/Pacific Islander"
\item "Black"
\item "Hispanic"
\item "White (includes Hispanic)"
\item "White non-Hispanic"
\item "Hispanic (Any Race)"
}}

\item{sex}{One of the following values
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Households, Rank
}
\description{
This function returns a data frame from population in State Cancer Profiles
}
\examples{
\dontrun{
demo_population(area = "WA",
                areatype = "county",
                population = "males",
                race = "all races (includes hispanic)")
                
demo_population(area = "dc",
                areatype = "hsa",
                population = "foreign born",
                race = "black",
                sex = "females")
                
demo_population(area = "usa",
                areatype = "state",
                population = "foreign born",
                race = "hispanic (any race)", 
                sex = "females")

}
}
