% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population.R
\name{demo_population}
\alias{demo_population}
\title{Access to Population Data}
\usage{
demo_population(area, areatype, population, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{population}{One of the following values: "age under 18", "age 18-39", "age 40-64", "ages 40 and over", "ages 50 and over", "ages 60 and over",
"american indian/alaska native", "asian/pacific islander", "black", "foreign born", "hispanic",
"non-hispanic (origin recode)", "white", "males", "females"}

\item{race}{One of the following values: "american indian/alaska native", "asian/pacific islander",
"black", "foreign born", "hispanic", "non-hispanic (origin recode)", "white"}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from population in State Cancer Profiles
}
\examples{
\dontrun{
demo_population("WA", "county", "asian/pacific islander", sex="females")
demo_population("dc", "hsa", "foreign born", "black", "females")
demo_population("usa", "state", "foreign born", "black", "females")

}
}
