% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-population.R
\name{demo_population}
\alias{demo_population}
\title{Access to Population Data}
\usage{
demo_population(area, areatype, population, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{population}{One of the following values:
\itemize{
\item \code{"age under 18"}
\item \code{"age 18-39"}
\item \code{"age 40-64"}
\item \code{"ages 40 and over"}
\item \code{"ages 50 and over"}
\item \code{"ages 60 and over"}
\item \code{"american indian/alaska native"}
\item \code{"asian/pacific islander"}
\item \code{"black"}
\item \code{"foreign born"}
\item \code{"hispanic"}
\item \code{"non-hispanic (origin recode)"}
\item \code{"white"}
\item \code{"males"}
\item \code{"females"}.
}}

\item{race}{One of the following values:
\itemize{
\item \code{"American Indian/Alaska Native"}
\item \code{"Asian/Pacific Islander"}
\item \code{"Black"}
\item \code{"Hispanic"}
\item \code{"White (includes Hispanic)"}
\item \code{"White non-Hispanic"}
\item \code{"Hispanic (Any Race)"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Households, Rank.
}
\description{
This function returns a data frame from population in State Cancer Profiles.
}
\examples{
\dontrun{
demo_population(area = "WA",
                areatype = "county",
                population = "males",
                race = "all races (includes hispanic)")
                
demo_population(area = "dc",
                areatype = "hsa",
                population = "foreign born",
                race = "black",
                sex = "females")
                
demo_population(area = "usa",
                areatype = "state",
                population = "foreign born",
                race = "hispanic (any race)", 
                sex = "females")

}
}
