% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-poverty.R
\name{demo_poverty}
\alias{demo_poverty}
\title{Access to Poverty Data}
\usage{
demo_poverty(area, areatype, poverty, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{poverty}{Either of the following values:
\itemize{
\item "families below poverty"
\item "persistent poverty"
\item "persons below poverty"
\item "persons < 150\% of poverty"
}}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "white (includes hispanic)"
\item "white non-hispanic"
\item "black"
\item "amer. indian/alaskan native (includes hispanic)"
\item "asian or pacific islander (includes hispanic)"
\item "hispanic (any race)"
}}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item Percent
\item Households
\item Rank
}
}
\description{
This function returns a data frame containing demographic data on Poverty in State Cancer Profiles.
}
\examples{
\dontrun{
demo_poverty(area = "WA",
             areatype = "county",
             poverty = "persistent poverty")
             
demo_poverty(area = "usa",
             areatype = "state",
             poverty = "families below poverty",
             race = "black")
             
demo_poverty(area = "dc",
             areatype = "hsa",
             poverty = "families below poverty",
             race = "All Races (includes Hispanic)")
}
}
