% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-poverty.R
\name{demo_poverty}
\alias{demo_poverty}
\title{Access to Poverty Data}
\usage{
demo_poverty(area, areatype, poverty, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{poverty}{One of the following values:
\itemize{
\item \code{"families below poverty"}
\item \code{"persistent poverty"}
\item \code{"persons below poverty"}
\item \code{"persons < 150\% of poverty"}.
}}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (includes Hispanic)"}
\item \code{"White non-Hispanic"}
\item \code{"Black"}
\item \code{"Amer. Indian/Alaskan Native (includes Hispanic)"}
\item \code{"Asian or Pacific Islander (includes Hispanic)"}
\item \verb{"Hispanic (Any Race)}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"males"}
\item \code{"females"}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Percent, Households, Rank.
}
\description{
This function returns a data frame about poverty demographics
from State Cancer Profiles.
}
\examples{
\dontrun{
demo_poverty(
  area = "WA",
  areatype = "county",
  poverty = "persistent poverty"
)

demo_poverty(
  area = "usa",
  areatype = "state",
  poverty = "families below poverty",
  race = "black"
)

demo_poverty(
  area = "dc",
  areatype = "hsa",
  poverty = "families below poverty",
  race = "All Races (includes Hispanic)"
)
}
}
\seealso{
Other demographics: 
\code{\link{demo_crowding}()},
\code{\link{demo_education}()},
\code{\link{demo_food}()},
\code{\link{demo_income}()},
\code{\link{demo_insurance}()},
\code{\link{demo_language}()},
\code{\link{demo_mobility}()},
\code{\link{demo_population}()},
\code{\link{demo_svi}()},
\code{\link{demo_workforce}()}
}
\concept{demographics}
