% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-poverty.R
\name{demo_poverty}
\alias{demo_poverty}
\title{Access to Poverty Data}
\usage{
demo_poverty(area, areatype, poverty, race = NULL, sex = NULL)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{poverty}{Either "families below poverty", "persistent poverty", "persons below poverty", "persons < 150\% of poverty"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "Households", "Rank"
}
\description{
This function returns a data frame from Poverty in State Cancer Profiles
}
\examples{
\dontrun{
demo_poverty(area = "WA",
             areatype = "county",
             poverty = "persistent poverty")
             
demo_poverty(area = "usa",
             areatype = "state",
             poverty = "families below poverty",
             race = "black")
             
demo_poverty(area = "dc",
             areatype = "hsa",
             poverty = "families below poverty",
             race = "All Races (includes Hispanic)")
}
}
