% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-svi.R
\name{demo_svi}
\alias{demo_svi}
\title{Access to Social Vulnerability Index (SVI) Data}
\usage{
demo_svi(area, svi)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{svi}{Either of the following values:
\itemize{
\item "Overall"
\item "socioeconomic status"
\item "household characteristics"
\item "racial & ethnic minority status"
\item "housing type & transportation"
}}
}
\value{
A data frame with the following columns "County", "FIPS", "Score"
}
\description{
This function returns a data frame containing demographic data on Social Vulnerability Index (SVI) in State Cancer Profiles.
}
\examples{
\dontrun{
demo_svi(area = "WA",
         svi = "overall")
         
demo_svi(area = "usa",
         svi = "overall")
         
demo_svi(area = "dc",
         svi = "socioeconomic status")
}
}
