% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-workforce.R
\name{demo_workforce}
\alias{demo_workforce}
\title{Access to Workforce Data}
\usage{
demo_workforce(area, areatype, race, sex)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Workforce in State Cancer Profiles
}
\examples{
\dontrun{
demo_workforce("WA", "county", "all races (includes hispanic)", "both sexes")
demo_workforce("usa", "state", "all races (includes hispanic)", "females")
demo_workforce("pr", "hsa", "all races (includes hispanic)", "both sexes")
}
}
