% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-workforce.R
\name{demo_workforce}
\alias{demo_workforce}
\title{Access to Workforce Data}
\usage{
demo_workforce(area, areatype, workforce, race, sex)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{One of the following values:
\itemize{
\item \code{"county"}
\item \code{"hsa"} (Health Service Area)
\item \code{"state"}.
}}

\item{workforce}{The only permissible value is
\code{"unemployed"}}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (includes Hispanic)"}
\item \code{"White non-Hispanic"}
\item \code{"Black"}
\item \code{"Amer. Indian/Alaskan Native (includes Hispanic)"}
\item \code{"Asian or Pacific Islander (includes Hispanic)"}
\item \verb{"Hispanic (Any Race)}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, People Unemployed, Rank.
}
\description{
This function returns a data frame from Workforce in State Cancer Profiles.
}
\examples{
\dontrun{
demo_workforce(area = "WA",
               areatype = "county",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
               
demo_workforce(area = "usa",
               areatype = "state",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "females")
               
demo_workforce(area = "pr",
               areatype = "hsa",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
}
}
