% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-workforce.R
\name{demo_workforce}
\alias{demo_workforce}
\title{Access to Workforce Data}
\usage{
demo_workforce(area, areatype, workforce, race, sex)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{workforce}{"unemployed"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Workforce in State Cancer Profiles
}
\examples{
\dontrun{
demo_workforce(area = "WA",
               areatype = "county",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
               
demo_workforce(area = "usa",
               areatype = "state",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "females")
               
demo_workforce(area = "pr",
               areatype = "hsa",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
}
}
