% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-workforce.R
\name{demo_workforce}
\alias{demo_workforce}
\title{Access to Workforce Data}
\usage{
demo_workforce(area, areatype, workforce, race, sex)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either \code{"county"}, \code{"hsa"} (Health Service Area), or
\code{"state"}.}

\item{workforce}{The only permissible values are
\code{"unemployed"}}

\item{race}{One of the following values
\itemize{
\item "All Races (includes Hispanic)"
\item "White (includes Hispanic)"
\item "White non-Hispanic"
\item "Black"
\item "Amer. Indian/Alaskan Native (includes Hispanic)"
\item "Asian or Pacific Islander (includes Hispanic)"
\item "hispanic (Any Race)
}}

\item{sex}{One of the following values
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, People Unemployed, Rank
}
\description{
This function returns a data frame from Workforce in State Cancer Profiles
}
\examples{
\dontrun{
demo_workforce(area = "WA",
               areatype = "county",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
               
demo_workforce(area = "usa",
               areatype = "state",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "females")
               
demo_workforce(area = "pr",
               areatype = "hsa",
               workforce = "unemployed",
               race = "all races (includes hispanic)",
               sex = "both sexes")
}
}
