% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food-access.R
\name{handle_food}
\alias{handle_food}
\title{Access to Food Insecurity Data}
\usage{
handle_food(food)
}
\arguments{
\item{food}{Either "food insecurity" or "limited access to healthy food"}

\item{area}{A state/territory abbreviation or USA.}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white non hispanic" = "01",
"black (includes hispanic)","hispanic (any race)}
}
\value{
A data frame with the following columns "County", "FIPS", "Value", "People"
}
\description{
This function returns a data frame from Food Insecurity in State Cancer Profiles
}
\examples{
demo_food("WA", "county", "food insecurity", "All Races (includes Hispanic)")
}
