% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-income.R
\name{handle_income}
\alias{handle_income}
\title{Handles Income Values to Code}
\usage{
handle_income(income)
}
\arguments{
\item{income}{Either "median family income" or "median household income"}
}
\value{
A string for its respective income
}
\description{
This function returns a matching code value for a Income for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_income("median family income")
}
}
