% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-mobility.R
\name{handle_mobility}
\alias{handle_mobility}
\title{Handles Mobility Values to Code}
\usage{
handle_mobility(mobility)
}
\arguments{
\item{mobility}{Either "i haven't moved (in past year)", "moved from outside us (in past year)",
"moved, different state (in past year)", "moved, different county, same state (in past year)",
"moved, same county (in past year)"}
}
\value{
A string for its respective mobility
}
\description{
This function returns a matching code value for a mobility for the api to use to get data from State Cancer Profiles
}
\examples{
\dontrun{
handle_mobility("i haven't moved (in past year)")
}
}
