% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle-race.R, R/workforce.R
\name{handle_race}
\alias{handle_race}
\title{Handles Race Values to Code}
\usage{
handle_race(race)

handle_race(race)
}
\arguments{
\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county" or "HSA" (Health service area)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A string for its respective race

A data frame with the following columns "County", "FIPS", "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a matching code value for a race for the api to use to get data from State Cancer Profiles

This function returns a data frame from Workforce in State Cancer Profiles
}
\examples{
handle_race("all races (includes hispanic)"), handle_race(race)
demo_workforce("WA", "county", "All Races (includes Hispanic)", "both sexes")
}
