% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svi.R
\name{handle_svi}
\alias{handle_svi}
\title{Access to Social Vulnerability Index (SVI) Data}
\usage{
handle_svi(svi)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{social}{Either "Overall, "socioeconomic status", "household characteristics", "racial & ethinic minority status", "housing type & transportation"}
}
\value{
A data frame with the following columns "County", "FIPS", "Score"
}
\description{
This function returns a data frame from Social Vulnerability Index (SVI) in State Cancer Profiles
}
\examples{
demo_svi("WA", "overall")
}
