% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence-cancer.R
\name{incidence_cancer}
\alias{incidence_cancer}
\title{Access to Cancer Incident Data}
\usage{
incidence_cancer(area, areatype, cancer, race, sex, age, stage, year)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{cancer}{Either "all cancer sites","bladder", "brain & ons", "breast (female)", "breast (female in situ)", "cervix",
"childhood (ages <15, all sites)", "childhood (ages <20, all sites)", "colon & rectum", "esophagus",
"kidney & renal pelvis", "leukemia", "liver & bile duct", "lung & bronchus", "melanoma of the skin",
"non-hodgkin lymphoma", "oral cavity & pharynx", "ovary", "pancreas", "prostate", "stomach",
"thyroid", "uterus (corpus & uterus, nos)"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{sex}{Either "both sexes", "males", "females"}

\item{age}{Either "all ages", "ages <50", "ages 50+", "ages <65", "ages 65+", ages <15, ages <20}

\item{stage}{Either "all stages" or "late stage (regional & distant)"}

\item{year}{Either "latest 5 year average", "latest single year (us by state)"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Age Adjusted Incidence Rate", "Lower 95\% CI",
"Upper 95\% CI", "CI Rank", "Lower CI Rank", "Upper CI Rank", "Annual Average Count", "Recent Trend",
"Recent 5 Year Trend", "Trend Lower 95\% CI", "Trend Upper 95\% CI"
}
\description{
This function returns a data frame from Cancer Incident in State Cancer Profiles
}
\examples{
\dontrun{
incidence_cancer(area = "wa",
                 areatype = "county",
                 cancer = "all cancer sites",
                 race = "black (non-hispanic)",
                 sex = "both sexes",
                 age = "ages 65+",
                 stage = "all stages",
                 year = "latest 5 year average")
                 
incidence_cancer(area = "usa",
                 areatype = "state",
                 cancer = "lung & bronchus",
                 race = "all races (includes hispanic)",
                 sex = "males",
                 age = "ages 50+",
                 stage = "late stage (regional & distant)",
                 year = "latest single year (us by state)")
                 
incidence_cancer(area = "wa",
                 areatype = "hsa",
                 cancer = "ovary",
                 race = "all races (includes hispanic)",
                 sex = "females",
                 age = "ages 50+",
                 stage = "late stage (regional & distant)",
                 year = "latest 5 year average")
}
}
