% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality-cancer.R
\name{mortality_cancer}
\alias{mortality_cancer}
\title{Access to Cancer Mortality Data}
\usage{
mortality_cancer(
  area,
  areatype,
  cancer,
  race,
  sex = NULL,
  age,
  year = "latest 5 year average"
)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{cancer}{Either "all cancer sites","bladder", "brain & ons", "breast (female)", "cervix",
"childhood (ages <15, all sites)", "childhood (ages <20, all sites)", "colon & rectum", "esophagus",
"kidney & renal pelvis", "leukemia", "liver & bile duct", "lung & bronchus", "melanoma of the skin",
"non-hodgkin lymphoma", "oral cavity & pharynx", "ovary", "pancreas", "prostate", "stomach",
"thyroid", "uterus (corpus & uterus, nos)"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{sex}{Either "both sexes", "males", "females"}

\item{age}{Either "all ages", "ages <50", "ages 50+", "ages <65", "ages 65+", ages <15, ages <20}

\item{year}{Either "latest 5 year average", "latest single year (us by state)"}
}
\value{
A data frame with the following columns "State", "FIPS", "Percent", "Lower 95\% CI", "Upper 95\% CI", "Number of Respondents"
}
\description{
This function returns a data frame from Cancer Mortality in State Cancer Profiles
}
\details{
This is the details section this is where the finer details of the function are explained
}
\examples{
\dontrun{
mortality_cancer("wa", "county", "all cancer sites", "black (non-hispanic)", 
                 "both sexes", "ages 65+")
mortality_cancer("ca", "hsa", "lung & bronchus", "all races (includes hispanic)", 
                 "males", "ages 50+")
mortality_cancer("usa", "state", "prostate", "all races (includes hispanic)", 
                 "males", "ages 50+")
mortality_cancer(area="wa", areatype="county", cancer="ovary", race="all races (includes hispanic)", 
                 sex="females", age="ages 50+")
mortality_cancer("ca", "hsa", "childhood (ages <20, all sites)", "all races (includes hispanic)", 
                 "males", "ages <20")
}
}
