% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality-cancer.R
\name{mortality_cancer}
\alias{mortality_cancer}
\title{Access to Cancer Mortality Data}
\usage{
mortality_cancer(area, areatype, cancer, race, sex, age, year)
}
\arguments{
\item{area}{A state/territory abbreviation or USA.}

\item{areatype}{Either "county", "hsa" (Health service area), or "state"}

\item{cancer}{Either "all cancer sites","bladder", "brain & ons", "breast (female)", "cervix",
"childhood (ages <15, all sites)", "childhood (ages <20, all sites)", "colon & rectum", "esophagus",
"kidney & renal pelvis", "leukemia", "liver & bile duct", "lung & bronchus", "melanoma of the skin",
"non-hodgkin lymphoma", "oral cavity & pharynx", "ovary", "pancreas", "prostate", "stomach",
"thyroid", "uterus (corpus & uterus, nos)"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{sex}{Defaulted to NULL. Specify either "both sexes", "males", "females" if necessary}

\item{age}{Either "all ages", "ages <50", "ages 50+", "ages <65", "ages 65+", ages <15, ages <20}

\item{year}{Either "latest 5 year average", "latest single year (us by state)"}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Met Healthy People Objective of ***?",
"Age Adjusted Death Rate", "Lower 95\% CI Rate", "Upper 95\% CI Rate",
"CI Rank", "Lower CI Rank", "Upper CI Rank", "Annual Average Count",
"Recent Trend", "Recent 5 Year Trend", "Lower 95\% CI Trend", "Upper 95\% CI Trend"
}
\description{
This function returns a data frame from Cancer Mortality in State Cancer Profiles
}
\details{
This is the details section this is where the finer details of the function are explained
}
\examples{
\dontrun{
mortality_cancer(area = "wa",
                 areatype = "county",
                 cancer = "all cancer sites",
                 race = "black (non-hispanic)",
                 sex = "both sexes",
                 age = "ages 65+",
                 year = "latest 5 year average")

mortality_cancer(area = "usa",
                 areatype = "state",
                 cancer = "prostate",
                 race = "all races (includes hispanic)",
                 sex = "males",
                 age = "ages 50+",
                 year = "latest single year (us by state)")
                 
mortality_cancer(area = "wa",
                 areatype = "hsa",
                 cancer = "ovary",
                 race = "all races (includes hispanic)",
                 sex = "females",
                 age = "ages 50+",
                 year = "latest 5 year average")
}
}
