% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-alcohol.R
\name{risk_alcohol}
\alias{risk_alcohol}
\title{Access to Alcohol Screening and Risk Data}
\usage{
risk_alcohol(alcohol, race, sex)
}
\arguments{
\item{alcohol}{The only permissible value is
\code{"binge drinking (4+ drinks on one occasion for women, 5+ drinks for one occasion for men), ages 21+"}.}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (non-Hispanic)"}
\item \code{"Black (non-Hispanic)"}
\item \code{"American Indian / Alaska Native (non-Hispanic)"}
\item \code{"Asian / Pacific Islander (non-Hispanic)"}
\item \code{"Hispanic (Any Race)"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"male"}
\item \code{"female"}.
}}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Percent, Lower 95\% CI,
Upper 95\% CI, Number of Respondents.
}
\description{
This function returns a data frame about alcohol risks
from State Cancer Profiles.
}
\examples{
\dontrun{
risk_alcohol(
  alcohol = paste(
    "binge drinking (4+ drinks on one occasion for women,",
    "5+ drinks for one occasion for men), ages 21+"
  ),
  race = "all races (includes hispanic)",
  sex = "both sexes"
)
}
}
