% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-alcohol.R
\name{risk_alcohol}
\alias{risk_alcohol}
\title{Access to Alcohol Screening and Risk Data}
\usage{
risk_alcohol(alcohol, race, sex)
}
\arguments{
\item{alcohol}{"binge drinking (4+ drinks on one occasion for women, 5+ drinks for one occasion for men), ages 21+"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns "State", "FIPS", "Percent", "Lower 95\% CI", "Upper 95\% CI", "Number of Respondents"
}
\description{
This function returns a data frame from Alcohol in State Cancer Profiles
}
\examples{
\dontrun{
risk_alcohol(paste("binge drinking (4+ drinks on one occasion for women,", 
                   "5+ drinks for one occasion for men), ages 21+"), 
                   "all races (includes hispanic)", "both sexes")
}
}
