% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-colorectal-screening.R
\name{risk_colorectal_screening}
\alias{risk_colorectal_screening}
\title{Access to Colorectal Screening Data}
\usage{
risk_colorectal_screening(screening, race = NULL, sex = NULL, area = NULL)
}
\arguments{
\item{screening}{One of the following values: "ever had fobt, ages 50-75",
"guidance sufficient crc, ages 50-75",
"had colonoscopy in past 10 years, ages 50-75"
"home blood stool test in the past year, ages 45-75"
"received at least one recommended crc test, ages 45-75"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns: Area Type, Area Code, "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Colorectal Screening in State Cancer Profiles
}
\examples{
\dontrun{
risk_colorectal_screening(screening = "home blood stool test in the past year, ages 45-75",
                          race = "all races (includes hispanic)",
                          sex = "both sexes")
                          
risk_colorectal_screening(screening = "ever had fobt, ages 50-75",
                          area="usa")
risk_colorectal_screening(screening = "received at least one recommended crc test, ages 45-75",
                          race = "all races (includes hispanic)",
                          sex = "both sexes")
}
}
