% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-colorectal-screening.R
\name{risk_colorectal_screening}
\alias{risk_colorectal_screening}
\title{Access to Colorectal Screening Data}
\usage{
risk_colorectal_screening(screening, race = NULL, sex = NULL, area = NULL)
}
\arguments{
\item{screening}{One of the following values:
\itemize{
\item \code{"ever had fobt, ages 50-75"}
\item \code{"guidance sufficient crc, ages 50-75"}
\item \code{"had colonoscopy in past 10 years, ages 50-75"}
\item \code{"home blood stool test in the past year, ages 45-75"}
\item \code{"received at least one recommended crc test, ages 45-75"}.
}}

\item{race}{One of the following values:
\itemize{
\item \code{"All Races (includes Hispanic)"}
\item \code{"White (non-Hispanic)"}
\item \code{"Black (non-Hispanic)"}
\item \code{"American Indian / Alaska Native (non-Hispanic)"}
\item \code{"Asian / Pacific Islander (non-Hispanic)"}
\item \code{"Hispanic (Any Race)"}.
}}

\item{sex}{One of the following values:
\itemize{
\item \code{"both sexes"}
\item \code{"males"}
\item \code{"females"}.
}}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns:
Area Type, Area Code, Percent, People Unemployed, Rank.
}
\description{
This function returns a data frame about
colorectal screening from State Cancer Profiles.
}
\examples{
\dontrun{
risk_colorectal_screening(
  screening = "home blood stool test in the past year, ages 45-75",
  race = "all races (includes hispanic)",
  sex = "both sexes"
)

risk_colorectal_screening(
  screening = "ever had fobt, ages 50-75",
  area = "usa"
)
risk_colorectal_screening(
  screening = "received at least one recommended crc test, ages 45-75",
  race = "all races (includes hispanic)",
  sex = "both sexes"
)
}
}
\seealso{
Other risks: 
\code{\link{risk_alcohol}()},
\code{\link{risk_diet_exercise}()},
\code{\link{risk_smoking}()},
\code{\link{risk_vaccines}()},
\code{\link{risk_women_health}()}
}
\concept{risks}
