% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-colorectal-screening.R
\name{risk_colorectal_screening}
\alias{risk_colorectal_screening}
\title{Access to Colorectal Screening Data}
\usage{
risk_colorectal_screening(screening, race = NULL, sex = NULL, area = NULL)
}
\arguments{
\item{screening}{One of the following values: "ever had colorectal endoscopy (sigmoidoscopy or colonoscopy), ages 50+",
"fobt (1yr) / flex sig (5yr) / fobt (3yr) / colonoscopy (10yr), ages 50-75",
"home-based fecal occult blood test (fobt) in past two years",
"home-based fobt in the past two years or ever had a colorectal endoscopy", "ever had fobt, ages 50-75",
"guidance sufficient crc, ages 50-75", "had colonoscopy in past 10 years, ages 50-75"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns "County", "FIPS", "Percent", "People Unemployed", "Rank"
}
\description{
This function returns a data frame from Colorectal Screening in State Cancer Profiles
}
\examples{
\dontrun{
risk_colorectal_screening("ever had colorectal endoscopy (sigmoidoscopy or colonoscopy), ages 50+",
                          "all races (includes hispanic)","both sexes")
risk_colorectal_screening("ever had fobt, ages 50-75", area="usa")
risk_colorectal_screening("ever had fobt, ages 50-75", area="wa")
risk_colorectal_screening("fobt (1yr) / flex sig (5yr) / fobt (3yr) / colonoscopy (10yr), ages 50-75", 
                          "all races (includes hispanic)", "males" )
}
}
