% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-diet-exercise.R
\name{risk_diet_exercise}
\alias{risk_diet_exercise}
\title{Access to Diet & Exercise Screening Data}
\usage{
risk_diet_exercise(diet_exercise, race, sex)
}
\arguments{
\item{diet_exercise}{The only permissible values are
\itemize{
\item \code{"bmi is healthy, ages 20+"}
\item \code{"bmi is obese, ages 20+"}
\item \code{"bmi is obese, high school survey"}
\item \code{"bmi is overweight, high school survey"}
\item \code{"consumed 1 or more fruits per day"}
\item \code{"consumed 1 or more vegetables per day"}
\item \code{"no leisure time physical activity"}
}}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "White (non-Hispanic)"
\item "Black (non-Hispanic)"
\item "American Indian / Alaska Native (non-Hispanic)"
\item "Asian / Pacific Islander (non-Hispanic)"
\item "Hispanic (Any Race)"
}}

\item{sex}{One of the following values
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Lower 95\% CI, Upper 95\% CI, Number of Respondents
}
\description{
This function returns a data frame about Diet & Exercise risk from State Cancer Profiles.
}
\examples{
\dontrun{
risk_diet_exercise(diet_exercise = "bmi is healthy, ages 20+",
                   race = "all races (includes hispanic)", 
                   sex = "both sexes")
risk_diet_exercise(diet_exercise = "bmi is obese, high school survey",
                   race = "all races (includes hispanic)",
                   sex = "males")
}
}
