% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-diet-exercise.R
\name{risk_diet_exercise}
\alias{risk_diet_exercise}
\title{Access to Diet & Exercise Screening Data}
\usage{
risk_diet_exercise(diet_exercise, race, sex)
}
\arguments{
\item{diet_exercise}{Either "bmi is healthy, ages 20+", "bmi is obese, ages 20+",
"bmi is obese, high school survey", "bmi is overweight, high school survey",
"consumed 1 or more fruits per day", "consumed 1 or more vegetables per day",
"no leisure time physical activity"}

\item{race}{One of the following values: "All Races (includes Hispanic)", "white (includes hispanic)" = "01",
"white non-hispanic","black","amer. indian/alaskan native (includes hispanic)",
"asian or pacific islander (includes hispanic)","hispanic (any race)}

\item{sex}{Either "both sexes", "male", "female"}
}
\value{
A data frame with the following columns "State", "FIPS", "Percent", "Lower 95\% CI", "Upper 95\% CI", "Number of Respondents"
}
\description{
This function returns a data frame from Diet & Exercise in State Cancer Profiles
}
\examples{
\dontrun{
risk_diet_exercise("bmi is healthy, ages 20+", "all races (includes hispanic)", "both sexes")
risk_diet_exercise("bmi is obese, high school survey", "all races (includes hispanic)", "males")
}
}
