% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-smoking.R
\name{risk_smoking}
\alias{risk_smoking}
\title{Access to Smoking Data}
\usage{
risk_smoking(smoking, race = NULL, sex = NULL, datatype = NULL, area = NULL)
}
\arguments{
\item{smoking}{Either "smoking laws (any)", "smoking laws (bars)", "smoking laws (restaurants)",
"smoking laws (workplace)", "smoking laws (workplace; restaurant; & bar)",
"smokers (stopped for 1 day or longer)", "smoking not allowed at work (all people)",
"smoking not allowed in home (all people)", "smoking not allowed at work (current smokers)",
"smoking not allowed at work (former/never smokers)", "smoking not allowed in home (current smokers)",
"smoking not allowed in home (former/never smokers)", "former smoker; ages 18+",
"former smoker, quit 1 year+; ages 18+", "smokers (ever); ages 18+", "e-cigarette use; ages 18+",
"smokers (current); ages 18+"}

\item{race}{One of the following values: "all races (includes hispanic)", "white (non-hispanic)",
"black (non-hispanic)", "amer. indian / ak native (non-hispanic)",
"asian / pacific islander (non-hispanic)","hispanic (any race)"}

\item{sex}{Either "both sexes", "males", "females"}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns #"State", "FIPS", "Percent", "Lower CI 95\%", "Upper CI 95\%", "Number of Respondents"
}
\description{
This function returns a data frame from Smoking in State Cancer Profiles
}
\examples{
\dontrun{
risk_smoking("smoking laws (any)")
risk_smoking("smokers (stopped for 1 day or longer)", sex="both sexes", 
              datatype="county level modeled estimates", area="wa")
risk_smoking("smoking not allowed at work (current smokers)", sex="both sexes", 
              datatype="direct estimates")
risk_smoking("former smoker; ages 18+", sex="both sexes", 
              datatype="county level modeled estimates", area="ca")
risk_smoking("smokers (ever); ages 18+", race="hispanic (any race)", sex="both sexes", 
              datatype="direct estimates")
risk_smoking("smokers (current); ages 18+", race="all races (includes hispanic)", 
              sex="both sexes", datatype="county level modeled estimates", area="wa")
}
}
