% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-smoking.R
\name{risk_smoking}
\alias{risk_smoking}
\title{Access to Smoking Data}
\usage{
risk_smoking(smoking, race = NULL, sex = NULL, datatype = NULL, area = NULL)
}
\arguments{
\item{smoking}{The only permissible values are
\itemize{
\item \code{"smoking laws (any)"}
\item \code{"smoking laws (bars)"}
\item \code{"smoking laws (restaurants)"}
\item \code{"smoking laws (workplace)"}
\item \code{"smoking laws (workplace; restaurant; & bar)"}
\item \code{"smokers (stopped for 1 day or longer)"}
\item \code{"smoking not allowed at work (all people)"}
\item \code{"smoking not allowed in home (all people)"}
\item \code{"smoking not allowed at work (current smokers)"}
\item \code{"smoking not allowed at work (former/never smokers)"}
\item \code{"smoking not allowed in home (current smokers)"}
\item \code{"smoking not allowed in home (former/never smokers)"}
\item \code{"former smoker; ages 18+"}
\item \code{"former smoker, quit 1 year+; ages 18+"}
\item \code{"smokers (ever); ages 18+"}
\item \code{"e-cigarette use; ages 18+"}
\item \code{"smokers (current); ages 18+"}
}}

\item{race}{One of the following values:
\itemize{
\item "All Races (includes Hispanic)"
\item "White (non-Hispanic)"
\item "Black (non-Hispanic)"
\item "American Indian / Alaska Native (non-Hispanic)"
\item "Asian / Pacific Islander (non-Hispanic)"
\item "Hispanic (Any Race)"
}}

\item{sex}{One of the following values
\itemize{
\item "both sexes"
\item "male"
\item "female"
}}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, Lower CI 95\%, Upper CI 95\%, Number of Respondents
}
\description{
This function returns a data frame about Smoking risks from State Cancer Profiles.
}
\details{
Please note that this function requires very specific arguments for each smoking type
}
\examples{
\dontrun{
risk_smoking(smoking = "smoking laws (any)")

risk_smoking(smoking = "smokers (stopped for 1 day or longer)",
             sex = "both sexes", 
             datatype = "county level modeled estimates",
             area = "wa")
             
risk_smoking(smoking = "smoking not allowed at work (current smokers)",
             sex = "both sexes", 
             datatype = "direct estimates")
             
risk_smoking(smoking = "smokers (current); ages 18+",
             race = "all races (includes hispanic)",
             sex = "both sexes",
             datatype = "county level modeled estimates",
             area="wa")
}
}
