% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-vaccines.R
\name{risk_vaccines}
\alias{risk_vaccines}
\title{Access to Vaccines Data}
\usage{
risk_vaccines(vaccine, sex)
}
\arguments{
\item{vaccine}{One of the following values:
\itemize{
\item "percent with up to date hpv vaccination coverage, ages 13-15"
\item "percent with up to date hpv vaccination coverage, ages 13-17"
}}

\item{sex}{Either "both sexes", "males", "females"}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item "Percent"
\item "Lower 95\% CI"
\item "Upper 95\% CI"
\item "Number of Respondents"
}
}
\description{
This function returns a data frame that includes data on screening and risk
factors related to vaccines and exercise as presented in the State Cancer Profiles.
}
\examples{
\dontrun{
risk_vaccines(vaccine = "percent with up to date hpv vaccination coverage, ages 13-15",
              sex = "both sexes")
              
risk_vaccines(vaccine = "percent with up to date hpv vaccination coverage, ages 13-17",
              sex = "females")
}
}
