% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-womens-health.R
\name{risk_women_health}
\alias{risk_women_health}
\title{Access to Women's Health Data}
\usage{
risk_women_health(
  women_health,
  race,
  datatype = "direct estimates",
  area = NULL
)
}
\arguments{
\item{women_health}{One of the following values
\itemize{
\item \code{"mammogram in past 2 years, ages 50-74"}
\item \code{"mammogram in past 2 years, ages 40+"}
\item \code{"pap smear in past 3 years, no hysterectomy, ages 21-65"}
}}

\item{race}{One of the following values
\itemize{
\item "All Races (includes Hispanic)"
\item "White (non-Hispanic)"
\item "Black (non-Hispanic)"
\item "American Indian / Alaska Native (non-Hispanic)"
\item "Asian / Pacific Islander (non-Hispanic)"
\item "Hispanic (Any Race)"
}}

\item{datatype}{One of the following values
\itemize{
\item \code{"direct estimates"}
\item \code{"county level modeled estimates"}
}}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns: Area Type, Area Code, Percent, People Unemployed, Rank
}
\description{
This function returns a data frame about women's health risks from State Cancer Profiles.
}
\examples{
\dontrun{
risk_women_health(women_health = "mammogram in past 2 years, ages 50-74", 
                  race = "all races (includes hispanic)",
                  datatype = "direct estimates")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysterectomy, ages 21-65", 
                  race = "all races (includes hispanic)",
                  datatype = "county level modeled estimates",
                  area = "wa")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysteroetomy, ages 21-65",
                  race = "black (non-hispanic)")
}
}
