% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-womens-health.R
\name{risk_women_health}
\alias{risk_women_health}
\title{Access to Women's Health Data}
\usage{
risk_women_health(
  women_health,
  race,
  datatype = "direct estimates",
  area = NULL
)
}
\arguments{
\item{women_health}{One of the following values:
\itemize{
\item "mammogram in past 2 years, ages 50-74"
\item "mammogram in past 2 years, ages 40+"
\item "pap smear in past 3 years, no hysterectomy, ages 21-65"
}}

\item{race}{One of the following values:
\itemize{
\item "all races (includes hispanic)"
\item "white (non-hispanic)"
\item "black (non-hispanic)"
\item "amer. indian / ak native (non-hispanic)"
\item "asian / pacific islander (non-hispanic)"
\item "hispanic (any race)"
}}

\item{datatype}{Either "direct estimates" or "county level modeled estimates"}

\item{area}{A state/territory abbreviation or USA.}
}
\value{
A data frame with the following columns:
\itemize{
\item Area Type
\item Area Code
\item "Percent"
\item "People Unemployed"
\item "Rank"
}
}
\description{
This function returns a data frame that includes data on screening and risk
factors related to women's health as presented in the State Cancer Profiles.
}
\examples{
\dontrun{
risk_women_health(women_health = "mammogram in past 2 years, ages 50-74", 
                  race = "all races (includes hispanic)",
                  datatype = "direct estimates")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysterectomy, ages 21-65", 
                  race = "all races (includes hispanic)",
                  datatype = "county level modeled estimates",
                  area = "wa")
                  
risk_women_health(women_health = "pap smear in past 3 years, no hysteroetomy, ages 21-65",
                  race = "black (non-hispanic)")
}
}
